+++ Data/IApplicantMessagesRespository.cs
@@ -5,7 +5,7 @@ namespace BankTrackingSystem.Data
     public interface IApplicantMessagesRespository
     {
         Task<List<ApplicantMessagesModel>> GetAllMessagesAsync();
-        Task<List<ApplicantMessagesModel>> GetAllAgainstApplicantId(int applicantId);
+        Task<List<ApplicantMessagesModel>> GetAllAgainstApplicantId(long applicantId);
         Task<ApplicantMessagesModel?> GetMessageAgainstIdAsync(int messageId);
         Task<ApplicantMessagesModel> AddMessage(ApplicantMessagesModel message);
     }
